Update local runner

+++ src/satori_cli/utils/execution/runner.py
@@ -8,8 +8,6 @@ from itertools import groupby
 from time import perf_counter
 from typing import Iterable
 
-from async_timeout import timeout_at
-
 from .models import CommandData, CommandLine, FileBasedResultCache, Result, ResultCache
 from .utils import replace_results, replace_testcase
 
@@ -18,7 +16,15 @@ log.setLevel(logging.INFO)
 log.addHandler(logging.StreamHandler())
 
 
-async def run_command(command: str, shell: bool | None = None) -> Result:
+timeout_event = asyncio.Event()
+
+
+async def run_command(
+    stop_event: asyncio.Event,
+    command_timeout_event: asyncio.Event,
+    command: str,
+    shell: bool | None = None,
+) -> Result:
     buffer_limit = 1024 * 1024 * 10
 
     try:
@@ -47,14 +53,25 @@ async def run_command(command: str, shell: bool | None = None) -> Result:
         }
 
     start = perf_counter()
+    p_task = asyncio.create_task(p.wait())
 
-    try:
-        return_code = await p.wait()
-    except asyncio.CancelledError:
+    tasks: list[asyncio.Task] = [
+        p_task,
+        asyncio.create_task(timeout_event.wait()),
+        asyncio.create_task(command_timeout_event.wait()),
+        asyncio.create_task(stop_event.wait()),
+    ]
+
+    done, pending = await asyncio.wait(tasks, return_when="FIRST_COMPLETED")
+
+    if done != {p_task}:
         if not shell:
             p.kill()
         else:
             os.killpg(p.pid, signal.SIGKILL)
+
+    for t in done | pending:
+        t.cancel()
 
     stdout, stderr = await p.communicate()
     return_code = await p.wait()
@@ -69,23 +86,38 @@ async def run_command(command: str, shell: bool | None = None) -> Result:
     }
 
 
+async def set_after(event: asyncio.Event, after: float):
+    await asyncio.sleep(after)
+    event.set()
+
+
+class TimedOut(Exception): ...
+
+
+class Stopped(Exception): ...
+
+
 async def process_commands(
     command_lines: Iterable[CommandLine],
     commands_data: dict[str, CommandData],
     timeout: int | None = None,
     cache_class: type[ResultCache] = FileBasedResultCache,
+    stop_event: asyncio.Event | None = None,
 ):
-    loop = asyncio.get_running_loop()
-    deadline = loop.time() + int(timeout) if timeout is not None else None
+    if not stop_event:
+        stop_event = asyncio.Event()
 
-    def to_deadline(timeout: int | None):
-        if timeout:
-            new_deadline = loop.time() + timeout
-            return min(deadline, new_deadline) if deadline else new_deadline
-        return deadline
+    if timeout is not None:
+        asyncio.create_task(set_after(timeout_event, timeout))
 
     grouped = groupby(command_lines, lambda x: x["path"])
     grouped_data = ((commands_data[path], clines) for path, clines in grouped)
+
+    def check_events():
+        if stop_event.is_set():
+            raise Stopped
+        if timeout_event.is_set():
+            raise TimedOut
 
     with cache_class() as result_cache:
 
@@ -95,6 +127,8 @@ async def process_commands(
             )
 
         for data, clgroup in grouped_data:
+            check_events()
+
             inner_timeout = data["settings"].get("setCommandTimeout")
             parallel = data["settings"].get("setParallel")
             shell = data["settings"].get("setShell")
@@ -105,19 +139,22 @@ async def process_commands(
                 add_command = ([(cl, build_command(cl))] for cl in clgroup)
 
             for group_to_run in add_command:
+                command_timeout_event = asyncio.Event()
+
                 tasks: dict[asyncio.Task[Result], CommandLine] = {
-                    asyncio.create_task(run_command(command, shell)): cline
+                    asyncio.create_task(
+                        run_command(stop_event, command_timeout_event, command, shell)
+                    ): cline
                     for cline, command in group_to_run
                 }
 
                 for _, v in tasks.items():
                     log.info(f"Running {v['path']}: {v['original']}")
 
-                try:
-                    async with timeout_at(to_deadline(inner_timeout)):
-                        await asyncio.gather(*tasks)
-                except TimeoutError:
-                    pass
+                if inner_timeout is not None:
+                    asyncio.create_task(set_after(command_timeout_event, inner_timeout))
+
+                await asyncio.gather(*tasks)
 
                 for task, command_line in tasks.items():
                     result = await task
@@ -127,5 +164,4 @@ async def process_commands(
 
                     yield command_line, result
 
-                if deadline and deadline < loop.time():
-                    raise TimeoutError
+                check_events()
